Extracted vehicle rental options into separate object
